feat(schemas):  move pydantic model to schemas directory
